chore: moved schema types to types dir
